+++ packages/pangea-node-sdk/src/request.ts
@@ -9,13 +9,7 @@ import * as qs from "neoqs/legacy";
 import PangeaConfig, { version } from "./config.js";
 import { PangeaErrors } from "./errors.js";
 import { AttachedFile, PangeaResponse } from "./response.js";
-import {
-  ConfigEnv,
-  FileData,
-  FileItems,
-  PostOptions,
-  TransferMethod,
-} from "./types.js";
+import { FileData, FileItems, PostOptions, TransferMethod } from "./types.js";
 import { getHeaderField } from "./utils/multipart.js";
 
 const delay = async (ms: number) =>
@@ -163,11 +157,11 @@ class PangeaRequest {
     return getHeaderField(contentDisposition, "filename", null);
   }
 
-  private getFilenameFromURL(url: string): string | undefined {
-    return new URL(url).pathname.split("/").pop();
+  private getFilenameFromURL(url: URL): string | undefined {
+    return url.pathname.split("/").pop();
   }
 
-  public async downloadFile(url: string): Promise<AttachedFile> {
+  public async downloadFile(url: URL): Promise<AttachedFile> {
     const response = await this.httpRequest(url, {
       method: "GET",
       retry: { limit: this.config.requestRetries },
@@ -265,7 +259,7 @@ class PangeaRequest {
     const presigned_url = response.accepted_result.post_url;
     const file_details = response.accepted_result?.post_form_data;
 
-    this.postPresignedURL(presigned_url, {
+    this.postPresignedURL(new URL(presigned_url), {
       file: fileData.file,
       file_details: file_details,
       name: "file",
@@ -273,10 +267,7 @@ class PangeaRequest {
     return response;
   }
 
-  public async postPresignedURL(
-    url: string,
-    fileData: FileData
-  ): Promise<void> {
+  public async postPresignedURL(url: URL, fileData: FileData): Promise<void> {
     if (!fileData.file_details) {
       throw new PangeaErrors.PangeaError(
         "file_details should be defined to do a post"
@@ -306,7 +297,7 @@ class PangeaRequest {
     }
   }
 
-  public async putPresignedURL(url: string, fileData: FileData): Promise<void> {
+  public async putPresignedURL(url: URL, fileData: FileData): Promise<void> {
     if (fileData.file_details) {
       throw new PangeaErrors.PangeaError(
         "file_details should be undefined to do a put"
@@ -402,7 +393,7 @@ class PangeaRequest {
 
   /** Wrapper around `fetch()` POST with got-like options. */
   private async httpPost(
-    url: string,
+    url: URL,
     options: {
       body: AsyncIterable<Uint8Array> | FormData | string;
       headers?: Record<string, string>;
@@ -429,7 +420,7 @@ class PangeaRequest {
   }
 
   private async httpRequest(
-    url: string,
+    url: URL,
     options: {
       body?: AsyncIterable<Uint8Array> | Buffer | FormData | string;
       headers?: Record<string, string>;
@@ -446,6 +437,8 @@ class PangeaRequest {
     const fetchOptions: RequestInit = {
       duplex: "half",
       method: options.method,
+      // @ts-expect-error difference in `FormData` types between undici-types
+      // and formdata-node.
       body: options.body,
       headers: options.headers,
     };
@@ -598,24 +591,11 @@ class PangeaRequest {
     }
   }
 
-  public getUrl(path: string): string {
-    let url: string;
-    if (
-      this.config.domain.startsWith("http://") ||
-      this.config.domain.startsWith("https://")
-    ) {
-      url = `${this.config.domain}/${path}`;
-    } else {
-      const schema = this.config?.insecure === true ? "http://" : "https://";
-
-      if (this.config?.environment === ConfigEnv.LOCAL) {
-        url = `${schema}${this.config.domain}/${path}`;
-      } else {
-        url = `${schema}${this.serviceName}.${this.config.domain}/${path}`;
-      }
-    }
-
-    return url;
+  public getUrl(path: string): URL {
+    return new URL(
+      path,
+      this.config.baseUrlTemplate.replace("{SERVICE_NAME}", this.serviceName)
+    );
   }
 
   private getHeaders(): Record<string, string> {
@@ -648,7 +628,7 @@ class PangeaRequest {
       case "ValidationError":
         throw new PangeaErrors.ValidationError(response.summary, response);
       case "TooManyRequests":
-        throw new PangeaErrors.RateLimiteError(response.summary, response);
+        throw new PangeaErrors.RateLimitError(response.summary, response);
       case "NoCredit":
         throw new PangeaErrors.NoCreditError(response.summary, response);
       case "Unauthorized":