+++ packages/pangea-node-sdk/src/services/file_scan.ts
@@ -151,7 +151,7 @@ export class FileScanUploader {
 
   // TODO: Docs
   public async uploadFile(
-    url: string,
+    url: URL,
     fileData: FileData,
     options: {
       transfer_method?: TransferMethod;